refactor(eventstore): migrate to iter.Seq2 and int64 types
Change QueryEvents return type from iter.Seq[nostr.Event] to
iter.Seq2[nostr.Event, error], CountEvents return type from uint32
to int64, and Close() to Close() error across all backends.

Add CollectEvents and EventsOnly helpers for Seq2→Seq compatibility.
Update all backends (badgerdb, boltdb, pebbledb, nullstore, slicestore),
relay layer, wrappers, and tests to match the new interface.

Also add experimental PebbleDB2 backend and related benchmarks/tests.

+++ eventstore/badgerdb/count.go
@@ -13,7 +13,7 @@ import (
 	"github.com/dgraph-io/badger/v4"
 )
 
-func (b *BadgerBackend) CountEvents(ctx context.Context, filter nostr.Filter) (uint32, error) {
+func (b *BadgerBackend) CountEvents(ctx context.Context, filter nostr.Filter) (int64, error) {
 	var count int = 0
 
 	queries, extraAuthors, extraKinds, extraTagKey, extraTagValues, since, e := b.prepareQueries(filter)
@@ -124,5 +124,5 @@ func (b *BadgerBackend) CountEvents(ctx context.Context, filter nostr.Filter) (u
 		return 0, fmt.Errorf("failed to count events: %w", e)
 	}
 
-	return uint32(count), nil
+	return int64(count), nil
 }
